// src/module_bindings/types.ts
// THIS FILE IS AUTOMATICALLY GENERATED BY SPACETIMEDB. EDITS TO THIS FILE
// WILL NOT BE SAVED. MODIFY TABLES IN YOUR MODULE SOURCE CODE INSTEAD.

/* eslint-disable */
/* tslint:disable */
import {
  TypeBuilder as __TypeBuilder,
  t as __t,
  type AlgebraicTypeType as __AlgebraicTypeType,
  type Infer as __Infer,
} from "spacetimedb";

export const GameEvent = __t.object("GameEvent", {
  id: __t.u64(),
  roomId: __t.u64(),
  senderIdentity: __t.identity(),
  eventType: __t.string(),
  payloadJson: __t.string(),
});
export type GameEvent = __Infer<typeof GameEvent>;

export const Player = __t.object("Player", {
  identity: __t.identity(),
  username: __t.string(),
  online: __t.bool(),
  wins: __t.u32(),
  coins: __t.u32(),
});
export type Player = __Infer<typeof Player>;

export const PlayerState = __t.object("PlayerState", {
  identity: __t.identity(),
  roomId: __t.u64(),
  x: __t.f64(),
  y: __t.f64(),
  angle: __t.f64(),
  score: __t.u32(),
  alive: __t.bool(),
  segmentsCount: __t.u32(),
  boosting: __t.bool(),
  skinJson: __t.string(),
  name: __t.string(),
});
export type PlayerState = __Infer<typeof PlayerState>;

export const Room = __t.object("Room", {
  id: __t.u64(),
  slug: __t.string(),
  name: __t.string(),
  host: __t.identity(),
  isPublic: __t.bool(),
  gameMode: __t.string(),
  maxPlayers: __t.u32(),
  seed: __t.u32(),
  createdAt: __t.timestamp(),
});
export type Room = __Infer<typeof Room>;

export const RoomMember = __t.object("RoomMember", {
  id: __t.u64(),
  roomId: __t.u64(),
  identity: __t.identity(),
  username: __t.string(),
  eliminated: __t.bool(),
});
export type RoomMember = __Infer<typeof RoomMember>;